+++ @utahdts/utah-design-system/react/components/forms/Switch.jsx
@@ -67,7 +67,6 @@ export function Switch({
         setInternalState(!!value);
       }
     },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     [value]
   );