+++ utah-design-system-website/src/react/hooks/useTextAreaCaretRowColumn.js
@@ -1,6 +1,5 @@
-import { useRefAlways } from '@utahdts/utah-design-system';
+import { useInterval, useRefAlways } from '@utahdts/utah-design-system';
 import { useState } from 'react';
-import { useInterval } from './useInterval';
 
 /**
  * This ONLY works if line wrapping is turned off because it counts line endings. Otherwise, soft wraps throw off the